requirements.txt
numpy==1.23.1
Pillow==9.2.0
opencv-python==4.6.0.66
imutils==0.5.4